Move TokenType into its own header (token_type.hpp)
Extract the TokenType enum out of the heavy parse_result.hpp into a lightweight duckdb/parser/peg/token_type.hpp so it can be included from the parser-extension API header without pulling in the transformer. Rename the ERROR member to TOKEN_ERROR to avoid colliding with the Win32 ERROR macro (the enum previously relied on parse_result.hpp's windows_undefs.hpp ordering).
