+++ tools/shell/linenoise/highlighting.cpp
@@ -28,7 +28,7 @@ static tokenType convertToken(TokenType token_type) {
 		return tokenType::TOKEN_KEYWORD;
 	case TokenType::COMMENT:
 		return tokenType::TOKEN_COMMENT;
-	case TokenType::ERROR:
+	case TokenType::TOKEN_ERROR:
 		return tokenType::TOKEN_ERROR;
 	case TokenType::END_OF_INPUT:
 	case TokenType::END_OF_INPUT_AUTOCOMPLETE:
